Move GroupedItesList to another file
